Extract AndyAB public api methods
I want users to clearly be able to see the methods available to them.
This is an initial attempt at extracting those public methods into the
AndyAB class.

NB: Extracting a class may increase the overall filesize of the library,
which is not good. I'll keep an eye on it and revert it if this is the
case.